wip(poly libs): get all third-party libraries that are used in a Polylith repo
