+++ components/polylith/libs/__init__.py
@@ -1,3 +1,4 @@
 from polylith.libs.grouping import get_third_party_imports
+from polylith.libs import report
 
-__all__ = ["get_third_party_imports"]
+__all__ = ["get_third_party_imports", "report"]
